+++ src/components/RoundedButton.tsx
@@ -2,15 +2,16 @@ import React from 'react'
 import { TouchableOpacity, Text, StyleSheet } from 'react-native'
 
 interface Props {
-    text: string
+    text: string,
+    onPress: () => void
 }
 
-export const RoundedButton = ({ text }: Props) => {
+export const RoundedButton = ({ text, onPress }: Props) => {
     return (
 
         <TouchableOpacity
             style={ styles.roundedButton }
-            onPress={() => {}}
+            onPress={() => onPress()}
         >
             <Text style={ styles.textButton }>{ text }</Text>
         </TouchableOpacity>
